moved test cases to a seperate file
